requirements.txt
Flask==3.0.3
Flask-Login==0.6.3
Flask-WTF==1.2.1
SQLAlchemy==2.0.32
email-validator==2.2.0
python-dotenv==1.0.1
Werkzeug==3.0.3
qrcode==7.4.2
pillow==10.4.0
gunicorn==21.2.0
psycopg[binary]==3.1.18
